Tweak the window tips examples to wrap better.
On my browser (chromium) in my window size the example was disappearing off
the edge of the screen. The browser had added a scroll bar, but I was
momentarily confused.

This tiny tweak formats it better, and adds a gap between the query and the
output, which makes it easier (at least for me) to comprehend.